fix: addressing meeting comments
